chore: update architecture rename application to modules
update exports index and imports on files
