+++ frontend/__tests__/HeaderLogo.test.tsx
@@ -2,12 +2,12 @@
 import { cleanup, render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
 
-import { HeaderLogo } from '@/app/application/components'
-import type { Children } from '@/app/application/schemas'
+import { HeaderLogo } from '@/app/modules/components/layout'
+import type { Children } from '@/app/modules/schemas'
 
 afterEach(cleanup)
 
-jest.mock('../src/app/application/components/LinkButton', () => ({
+jest.mock('../src/app/modules/components/LinkButton', () => ({
   __esModule: true,
   LinkButton: ({ children }: Children) => <a role='button' href='/'>{children} Mocked LinkButton Component</a>
 }))
