--- frontend/src/app/application/components/carousel/index.ts
@@ -1,2 +1,0 @@
-export * from './CarouselButton'
-export * from './TrackedCarousel'
