--- frontend/src/app/application/components/index.ts
@@ -1,7 +1,0 @@
-export * from './layout'
-export * from './templates'
-export * from './carousel'
-
-export * from './LinkButton'
-export * from './SVGs'
-export * from './Button'
