--- frontend/src/app/application/components/layout/index.ts
@@ -1,5 +1,0 @@
-export * from './HeaderLayout'
-export * from './HeaderLogo'
-export * from './headerPages.ref'
-export * from './FooterLayout'
-export * from './FooterLink'
