--- frontend/src/app/application/components/templates/index.ts
@@ -1,3 +1,0 @@
-export * from './Header'
-export * from './Footer'
-export * from './Navigation'
