--- frontend/src/app/application/hocs/index.ts
@@ -1,1 +1,0 @@
-export * from './withTrackerPosition.hoc'
