--- frontend/src/app/application/hooks/carousel/index.ts
@@ -1,1 +1,0 @@
-export * from './useCarouselEffect.hook'
