--- frontend/src/app/application/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './isValidArray.utility'
-export * from './generateRange.utility'
