+++ frontend/src/app/home/components/CoverCarousel.home.tsx
@@ -1,11 +1,10 @@
 'use client'
 
-import { CoverImage } from './CoverImage'
-import { TrackedCarousel } from '@/application/components/carousel'
-import { withTrackerPosition } from '@/application/hocs'
-
-import { SquareIcon as IconComponent } from '@/application/assets'
-import type { Slides } from './data.ref'
+import { TrackedCarousel } from "@/app/modules/components/carousel"
+import { withTrackerPosition } from "@/app/modules/hocs"
+import { SquareIcon as IconComponent } from "@/app/modules/assets"
+import { CoverImage } from "./CoverImage.home"
+import { Slides } from "../data.ref"
 
 const positionColor = {
   default: 'text-gray-900 fill-gray-900 hover:text-gray-800 hover:fill-gray-800',
