+++ frontend/src/app/home/components/CoverImage.home.tsx
