+++ frontend/src/app/home/components/index.ts
@@ -1,0 +1,3 @@
+// components
+export { CoverCarousel } from './CoverCarousel.home'
+export { CoverImage } from './CoverImage.home'
