--- frontend/src/app/home/index.ts
@@ -1,2 +1,0 @@
-export * from './CoverCarousel'
-export * from './CoverImage'
