+++ frontend/src/app/modules/assets/Icons.tsx
