+++ frontend/src/app/modules/assets/index.ts
