+++ frontend/src/app/modules/components/SVGs.tsx
