+++ frontend/src/app/modules/components/carousel/TrackedCarousel.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import { useCarouselEffect } from '@/application/hooks'
+import { useCarouselEffect } from '../../hooks'
 import { CarouselButton } from './CarouselButton'
-import { LeftArrowIcon, RightArrowIcon } from '@/application/assets'
+import { LeftArrowIcon, RightArrowIcon } from '../../assets'
 
-import type { BaseComponentProps } from '@/application/schemas'
-import type { EnhancedIconComponentProps } from '@/application/hocs'
+import type { EnhancedIconComponentProps } from '../../hocs'
+import type { BaseComponentProps } from '../../schemas'
 import type { AddReactComponent } from '@/utilities'
 
 const defaultBtnIcons = {
