+++ frontend/src/app/modules/components/carousel/index.ts
@@ -1,0 +1,3 @@
+// components
+export { CarouselButton } from './CarouselButton'
+export { TrackedCarousel } from './TrackedCarousel'
