+++ frontend/src/app/modules/components/index.ts
@@ -1,0 +1,12 @@
+// folders
+// export * from './layout'
+// export * from './templates'
+// export * from './carousel'
+
+// components
+export { LinkButton } from './LinkButton'
+export { SVGElement } from './SVGs'
+export { Button } from './Button'
+
+// types
+export type { LinkButtonProps } from './LinkButton'
