+++ frontend/src/app/modules/components/layout/HeaderLogo.tsx
