+++ frontend/src/app/modules/components/layout/index.ts
@@ -1,0 +1,13 @@
+// components
+export { HeaderLogo } from './HeaderLogo'
+export { FooterLink } from './FooterLink'
+export { HeaderLayout } from './HeaderLayout'
+export { FooterLayout } from './FooterLayout'
+
+// object
+export { headerPages } from './headerPages.ref'
+
+// types
+export type { HeaderLayoutProps } from './HeaderLayout'
+export type { FooterLinkProps } from './FooterLink'
+export type { HeaderLogoProps } from './HeaderLogo'
