+++ frontend/src/app/modules/components/templates/Footer.tsx
