+++ frontend/src/app/modules/components/templates/index.ts
@@ -1,0 +1,4 @@
+// components
+export { Header } from './Header'
+export { Footer } from './Footer'
+export { Navigation } from './Navigation'
