+++ frontend/src/app/modules/hocs/index.ts
@@ -1,0 +1,5 @@
+// components
+export { withTrackerPosition } from './withTrackerPosition.hoc'
+
+// types
+export type { EnhancedIconComponentProps } from './withTrackerPosition.hoc'
