+++ frontend/src/app/modules/hooks/carousel/carouselControls.ts
