+++ frontend/src/app/modules/hooks/carousel/index.ts
@@ -1,0 +1,2 @@
+// hooks
+export { useCarouselEffect } from './useCarouselEffect.hook'
