+++ frontend/src/app/modules/hooks/carousel/useCarouselEffect.hook.ts
