+++ frontend/src/app/modules/hooks/index.ts
@@ -1,1 +1,2 @@
+// folders
 export * from './carousel'
