+++ frontend/src/app/modules/utils/generateRange.utility.ts
