+++ frontend/src/app/modules/utils/index.ts
@@ -1,0 +1,3 @@
+// utilities
+export { isValidArray } from './isValidArray.utility'
+export { generateRange } from './generateRange.utility'
