+++ frontend/src/app/modules/utils/isValidArray.utility.ts
