+++ frontend/tsconfig.json
@@ -30,7 +30,7 @@
       "@/*": [
         "./src/*",
         "./src/app/*",
-        "./src/app/application/*"
+        "./src/app/modules/*"
       ]
     }
   },
